requirements.txt
streamlit
google-generativeai
pyngrok
pdfkit
wkhtmltopdf
markdown
boto3
